Updated tests, comments; changed req.path to req.route
