+++ src/utils/saga/index.js
@@ -1,14 +1,3 @@
-/* import { counter } from './counter';
-import { pager } from './pager';
-import { errorToast, successToast, toast } from './toasts';
-
-export const helpers = {
-  counter,
-  pager,
-  errorToast,
-  successToast,
-  toast
-};*/
 import counterUtilities from './counter';
 import pageUtilities from './pager';
 import toastUtilities from './toasts';
